api/requirements.txt
email-validator==1.3.1
fastapi==0.89.1
PyJWT==2.6.0
python-decouple==3.7
uvicorn==0.20.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
passlib==1.7.4